fix renaming of repo and docker files
