+++ utils/rxiv-download-tool/.bumpversion.cfg
@@ -1,5 +1,5 @@
 [bumpversion]
-current_version = 0.1.0-dev
+current_version = 0.1.0-dev0
 commit = False
 tag = False
 parse = (?P<major>\d+)\.(?P<minor>\d+)\.(?P<patch>\d+)(\-(?P<release>[a-z]+)(?P<dev>\d+))?
@@ -26,4 +26,4 @@ replace = version = "{new_version}"
 
 [bumpversion:file:plugin.json]
 
-[bumpversion:file:src/polus/plugins/utils/rxiv_download/__init__.py]
+[bumpversion:file:src/polus/images/utils/rxiv_download/__init__.py]
