+++ utils/rxiv-download-tool/build-docker.sh
@@ -1,0 +1,4 @@
+#!/bin/bash
+
+version=$(<VERSION)
+docker build . -t polusai/rxiv-download-tool:${version}
