+++ utils/rxiv-download-tool/examples/arXiv_20231218_0.json
