+++ utils/rxiv-download-tool/plugin.json
@@ -1,18 +1,18 @@
 {
   "name": "Rxiv-Download",
-  "version": "0.1.0-dev",
+  "version": "0.1.0-dev0",
   "title": "Download Rxiv text data",
   "description": "This plugin allows to download data from Rxiv website.",
   "author": "Nick Schaub (nick.schaub@nih.gov), Hamdah Shafqat Abbasi (hamdahshafqat.abbasi@nih.gov)",
   "institution": "National Center for Advancing Translational Sciences, National Institutes of Health",
-  "repository": "https://github.com/PolusAI/polus-plugins",
+  "repository": "https://github.com/PolusAI/image-tools",
   "website": "https://ncats.nih.gov/preclinical/core/informatics",
   "citation": "",
-  "containerId": "polusai/rxiv-download-plugin:0.1.0-dev",
+  "containerId": "polusai/rxiv-download-tool:0.1.0-dev0",
   "baseCommand": [
     "python3",
     "-m",
-    "polus.plugins.utils.rxiv_download"
+    "polus.images.utils.rxiv_download"
   ],
   "inputs": {
     "rxiv": {
