+++ utils/rxiv-download-tool/src/polus/images/utils/rxiv_download/__init__.py
@@ -1,3 +1,3 @@
 """Rxiv Download Plugin."""
 
-__version__ = "0.1.0-dev"
+__version__ = "0.1.0-dev0"
