--- utils/rxiv-download/build-docker.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-
-version=$(<VERSION)
-docker build . -t polusai/rxiv-download-plugin:${version}
